Testing refinement (#161)
* Propose plan for testing refinement

* Audit complete

* First pass

* Rename test modules

* Round 2

* Second pass

* Plus type checking pass

* Remove conftest duplicate

* Adding one E2E smoke test with encryption enabled; restored refined test_pipeline_contracts.py - kept 8 tests for integration contracts

* Deleted plans folder - strategy captured in TESTING_STANDARDS.md, DOCUMENTATION_STANDARDS.md, and AGENTS.MD; added @pytest.mark.integration to test_error_propagation.py; fix docs for conftest location

---------
