+++ tests/e2e/test_full_pipeline.py
@@ -22,7 +22,7 @@ Each test:
 
 from __future__ import annotations
 
-import json
+import shutil
 import subprocess
 from collections.abc import Generator
 from pathlib import Path
@@ -43,45 +43,51 @@ class TestFullPipelineExecution:
         return Path(__file__).resolve().parent.parent.parent
 
     @pytest.fixture
-    def pipeline_input_file(self, project_root: Path) -> Generator[Path, None, None]:
-        """Create a test input Excel file in the project input directory."""
-        input_file = project_root / "input" / "e2e_test_clients.xlsx"
-        df = create_test_input_dataframe(num_clients=3)
-        df.to_excel(input_file, index=False, engine="openpyxl")
+    def e2e_workdir(
+        self, project_root: Path, tmp_path_factory: pytest.TempPathFactory
+    ) -> Generator[Path, None, None]:
+        """Create a temporary working directory within the project root.
 
-        yield input_file
+        Typst requires absolute paths relative to the project root for asset
+        resolution (QR codes, logos). This workdir is inside project root
+        to satisfy that constraint while maintaining isolation.
+        """
+        workdir = project_root / f"tmp_e2e_{tmp_path_factory.mktemp('e2e').name}"
+        workdir.mkdir(parents=True, exist_ok=True)
+        (workdir / "input").mkdir(exist_ok=True)
+        (workdir / "output").mkdir(exist_ok=True)
+
+        # Copy base config to workdir
+        config_dir = workdir / "config"
+        shutil.copytree(project_root / "config", config_dir)
+
+        yield workdir
 
         # Cleanup
-        if input_file.exists():
-            input_file.unlink()
+        if workdir.exists():
+            shutil.rmtree(workdir)
+
+    @pytest.fixture
+    def pipeline_input_file(self, e2e_workdir: Path) -> Path:
+        """Create a test input Excel file in the E2E workdir."""
+        input_file = e2e_workdir / "input" / "e2e_test_clients.xlsx"
+        df = create_test_input_dataframe(num_clients=3)
+        df.to_excel(input_file, index=False, engine="openpyxl")
+        return input_file
 
     def run_pipeline(
         self,
         input_file: Path,
         language: str,
         project_root: Path,
+        e2e_workdir: Path,
         config_overrides: dict | None = None,
     ) -> subprocess.CompletedProcess:
-        """Run the viper pipeline via subprocess.
+        """Run the viper pipeline via subprocess using isolated config/output."""
+        config_dir = e2e_workdir / "config"
 
-        Parameters
-        ----------
-        input_file : Path
-            Path to input Excel file
-        language : str
-            Language code ('en' or 'fr')
-        project_root : Path
-            Project root (used for output directory within project tree)
-        config_overrides : dict, optional
-            Config parameters to override before running pipeline
-
-        Returns
-        -------
-        subprocess.CompletedProcess
-            Result of pipeline execution
-        """
         if config_overrides:
-            config_path = project_root / "config" / "parameters.yaml"
+            config_path = config_dir / "parameters.yaml"
             with open(config_path) as f:
                 config = yaml.safe_load(f)
 
@@ -107,6 +113,10 @@ class TestFullPipelineExecution:
             language,
             "--input",
             str(input_file.parent),
+            "--output",
+            str(e2e_workdir / "output"),
+            "--config",
+            str(config_dir),
         ]
 
         result = subprocess.run(
@@ -115,26 +125,20 @@ class TestFullPipelineExecution:
         return result
 
     def test_full_pipeline_english(
-        self, tmp_path: Path, pipeline_input_file: Path, project_root: Path
+        self, pipeline_input_file: Path, project_root: Path, e2e_workdir: Path
     ) -> None:
-        """Test complete pipeline execution with English language.
-
-        Real-world significance:
-        - Core pipeline functionality must work for English input
-        - Verifies all 9 steps execute successfully
-        - Checks that per-client PDFs are created
-        """
-        # Disable encryption for core E2E test (tests basic functionality)
+        """Test complete pipeline execution with English language."""
+        # Disable encryption for core E2E test
         config_overrides = {"encryption": {"enabled": False}}
         result = self.run_pipeline(
-            pipeline_input_file, "en", project_root, config_overrides
+            pipeline_input_file, "en", project_root, e2e_workdir, config_overrides
         )
 
         assert result.returncode == 0, f"Pipeline failed: {result.stderr}"
         assert "Pipeline completed successfully" in result.stdout
 
-        # Verify output structure (in project output directory)
-        output_dir = project_root / "output"
+        # Verify output structure in E2E workdir
+        output_dir = e2e_workdir / "output"
         assert (output_dir / "artifacts").exists()
         assert (output_dir / "pdf_individual").exists()
 
@@ -142,27 +146,24 @@ class TestFullPipelineExecution:
         pdfs = list((output_dir / "pdf_individual").glob("en_notice_*.pdf"))
         assert len(pdfs) == 3, f"Expected 3 PDFs but found {len(pdfs)}"
 
-    def test_full_pipeline_french(
-        self, tmp_path: Path, pipeline_input_file: Path, project_root: Path
-    ) -> None:
-        """Test complete pipeline execution with French language.
+        # Bundling runs by default (bundle_size: 100 in parameters.yaml)
+        assert (output_dir / "pdf_combined").exists()
 
-        Real-world significance:
-        - Multilingual support must work for French input
-        - Templates, notices, and metadata must be in French
-        - Verifies language parameter is respected throughout pipeline
-        """
-        # Disable encryption for core E2E test (tests basic functionality)
+    def test_full_pipeline_french(
+        self, pipeline_input_file: Path, project_root: Path, e2e_workdir: Path
+    ) -> None:
+        """Test complete pipeline execution with French language."""
+        # Disable encryption for core E2E test
         config_overrides = {"encryption": {"enabled": False}}
         result = self.run_pipeline(
-            pipeline_input_file, "fr", project_root, config_overrides
+            pipeline_input_file, "fr", project_root, e2e_workdir, config_overrides
         )
 
         assert result.returncode == 0, f"Pipeline failed: {result.stderr}"
         assert "Pipeline completed successfully" in result.stdout
 
-        # Verify output structure (in project output directory)
-        output_dir = project_root / "output"
+        # Verify output structure in E2E workdir
+        output_dir = e2e_workdir / "output"
         assert (output_dir / "artifacts").exists()
         assert (output_dir / "pdf_individual").exists()
 
@@ -170,198 +171,31 @@ class TestFullPipelineExecution:
         pdfs = list((output_dir / "pdf_individual").glob("fr_notice_*.pdf"))
         assert len(pdfs) == 3, f"Expected 3 French PDFs but found {len(pdfs)}"
 
-    def test_pipeline_with_qr_disabled(
-        self, tmp_path: Path, pipeline_input_file: Path, project_root: Path
+    def test_full_pipeline_with_encryption(
+        self, pipeline_input_file: Path, project_root: Path, e2e_workdir: Path
     ) -> None:
-        """Test pipeline with QR code generation disabled.
+        """Test complete pipeline execution with encryption enabled (Steps 7 + 9).
 
         Real-world significance:
-        - QR codes are optional (controlled by config)
-        - Pipeline must skip QR generation when disabled
-        - Should complete faster without QR generation
+        - Encryption is a core optional feature used in production deliveries
+        - Steps 7 and 9 are never exercised by the baseline smoke tests
+        - Verifies that encrypted PDFs are created alongside unencrypted originals
+
+        Assertion: One _encrypted.pdf exists per client in pdf_individual/
         """
-        # Disable both QR and encryption for this test
-        config_overrides = {
-            "qr": {"enabled": False},
-            "encryption": {"enabled": False},
-        }
-        result = self.run_pipeline(
-            pipeline_input_file, "en", project_root, config_overrides
-        )
-
-        assert result.returncode == 0, f"Pipeline failed: {result.stderr}"
-        assert "Step 3: Generating QR codes" in result.stdout
-        assert "disabled" in result.stdout.lower() or "skipped" in result.stdout.lower()
-
-        # Verify PDFs still exist
-        output_dir = project_root / "output"
-        pdfs = list((output_dir / "pdf_individual").glob("en_notice_*.pdf"))
-        assert len(pdfs) == 3
-
-    def test_pipeline_with_encryption(
-        self, tmp_path: Path, pipeline_input_file: Path, project_root: Path
-    ) -> None:
-        """Test pipeline with PDF encryption enabled.
-
-        Real-world significance:
-        - Encryption protects sensitive student data in PDFs
-        - Each PDF is encrypted with a unique password based on client data
-        - Encrypted versions are created alongside original PDFs
-        """
-        # Enable encryption for this specific test
         config_overrides = {"encryption": {"enabled": True}}
         result = self.run_pipeline(
-            pipeline_input_file, "en", project_root, config_overrides
+            pipeline_input_file, "en", project_root, e2e_workdir, config_overrides
         )
 
         assert result.returncode == 0, f"Pipeline failed: {result.stderr}"
-        assert "Encryption" in result.stdout
-        assert "success: 3" in result.stdout
+        assert "Pipeline completed successfully" in result.stdout
 
-        # Verify both encrypted and non-encrypted PDFs exist
-        output_dir = project_root / "output"
+        output_dir = e2e_workdir / "output"
         encrypted_pdfs = list(
-            (output_dir / "pdf_individual").glob("en_notice_*_encrypted.pdf")
+            (output_dir / "pdf_individual").glob("*_encrypted.pdf")
         )
         assert len(encrypted_pdfs) == 3, (
             f"Expected 3 encrypted PDFs but found {len(encrypted_pdfs)}"
         )
 
-        # Non-encrypted versions should also exist (not removed by default)
-        all_pdfs = list((output_dir / "pdf_individual").glob("en_notice_*.pdf"))
-        assert len(all_pdfs) == 6, (
-            f"Expected 6 total PDFs (3 encrypted + 3 non-encrypted) but found {len(all_pdfs)}"
-        )
-
-    def test_pipeline_with_batching(
-        self, tmp_path: Path, pipeline_input_file: Path, project_root: Path
-    ) -> None:
-        """Test pipeline with PDF bundling enabled.
-
-        Real-world significance:
-        - Bundling groups individual PDFs into combined files
-        - Useful for organizing output by school or size
-        - Creates manifests for audit trails
-        """
-        # Temporarily enable bundling in config
-        config_path = project_root / "config" / "parameters.yaml"
-        with open(config_path) as f:
-            config = yaml.safe_load(f)
-        original_bundle_size = config.get("bundling", {}).get("bundle_size")
-        original_encryption = config.get("encryption", {}).get("enabled")
-
-        try:
-            # Disable encryption and enable bundling
-            config["encryption"]["enabled"] = False
-            config["bundling"]["bundle_size"] = 2
-            with open(config_path, "w") as f:
-                yaml.dump(config, f)
-
-            result = self.run_pipeline(pipeline_input_file, "en", project_root)
-
-            assert result.returncode == 0, f"Pipeline failed: {result.stderr}"
-            assert "Bundling" in result.stdout
-            assert (
-                "created" in result.stdout.lower() or "bundle" in result.stdout.lower()
-            )
-
-            # Verify bundled PDFs exist
-            output_dir = project_root / "output"
-            assert (output_dir / "pdf_combined").exists()
-            bundles = list((output_dir / "pdf_combined").glob("en_bundle_*.pdf"))
-            assert len(bundles) > 0, "Expected bundled PDFs to be created"
-
-            # Verify manifests exist
-            assert (output_dir / "metadata").exists()
-            manifests = list((output_dir / "metadata").glob("*_manifest.json"))
-            assert len(manifests) == len(bundles)
-        finally:
-            # Restore original config
-            config["bundling"]["bundle_size"] = original_bundle_size
-            config["encryption"]["enabled"] = original_encryption
-            with open(config_path, "w") as f:
-                yaml.dump(config, f)
-
-    def test_pipeline_minimal_input(self, tmp_path: Path, project_root: Path) -> None:
-        """Test pipeline with minimal input (1 client).
-
-        Real-world significance:
-        - Pipeline must handle edge case of single client
-        - Single-client PDFs must work correctly
-        - Minimal input helps debug issues
-        """
-        # Create minimal input file with 1 client in project input dir
-        input_file = project_root / "input" / "e2e_minimal_input.xlsx"
-        df = create_test_input_dataframe(num_clients=1)
-        df.to_excel(input_file, index=False, engine="openpyxl")
-
-        try:
-            # Disable encryption for this test
-            config_overrides = {"encryption": {"enabled": False}}
-            result = self.run_pipeline(input_file, "en", project_root, config_overrides)
-
-            assert result.returncode == 0, f"Pipeline failed: {result.stderr}"
-            assert "Pipeline completed successfully" in result.stdout
-
-            # Verify single PDF was created
-            output_dir = project_root / "output"
-            pdfs = list((output_dir / "pdf_individual").glob("en_notice_*.pdf"))
-            assert len(pdfs) == 1
-        finally:
-            # Cleanup input file
-            if input_file.exists():
-                input_file.unlink()
-
-    def test_pipeline_validates_output_artifacts(
-        self, tmp_path: Path, pipeline_input_file: Path, project_root: Path
-    ) -> None:
-        """Test that pipeline creates valid output artifacts.
-
-        Real-world significance:
-        - Pipeline produces JSON artifacts that are read by other steps
-        - Artifacts must have correct schema (format, required fields)
-        - JSON corruption would cause silent failures in downstream steps
-        """
-        # Disable encryption for this test
-        config_overrides = {"encryption": {"enabled": False}}
-        result = self.run_pipeline(
-            pipeline_input_file, "en", project_root, config_overrides
-        )
-
-        assert result.returncode == 0
-
-        # Find and validate the preprocessed artifact
-        output_dir = project_root / "output"
-        artifacts = list((output_dir / "artifacts").glob("preprocessed_clients_*.json"))
-        assert len(artifacts) >= 1, "Expected at least 1 preprocessed artifact"
-
-        artifact = artifacts[0]
-        with open(artifact) as f:
-            data = json.load(f)
-
-        # Validate artifact structure
-        assert "run_id" in data
-        assert "language" in data
-        assert data["language"] == "en"
-        assert "clients" in data
-        assert len(data["clients"]) == 3
-        assert "warnings" in data
-
-        # Validate each client record
-        for client in data["clients"]:
-            assert "sequence" in client
-            assert "client_id" in client
-            assert "person" in client
-            assert "school" in client
-            assert "board" in client
-            assert "contact" in client
-            assert "vaccines_due" in client
-
-    def test_placeholder_e2e_marker_applied(self) -> None:
-        """Placeholder test ensuring e2e marker is recognized by pytest.
-
-        Real-world significance:
-        - E2E tests are marked so they can be run separately
-        - Can run only E2E tests with: uv run pytest -m e2e
-        """
-        assert True
